Rename Monad\Free2 to Monad\Free
